Add/Subtract/Multiply/Divide() requires 2 parameters, + do not use recursion
The signature for Add, Subtract, Multiply, and Divide functions no
longer allows them to be called with a single argument. I now know that
variadic means 0 or more of that type. :)

The processing of variadic parameters no longer uses recursive calls,
and now uses a loop to build the result. This logic became cleaner with
the above adjustment to parameters (a and b always being fixed
parameters instead of referencing the first element of b as a slice).

Along the way, replace `errors.New(fmt.Printf())` with `fmt.Errorf()`.